affects issues #68 and #69

+++ include/analysis/AgentStdDev.hxx
@@ -16,7 +16,7 @@ class AgentStdDev : public AgentAnalysis
 {
 	std::string _attributeName;
 	std::vector<int> _numAgents;
-	std::vector< std::vector<int> > _values;
+	std::vector< std::vector<float> > _values;
 public:
 	AgentStdDev( const std::string & attributeName );
 	virtual ~AgentStdDev();
